scripts/requirements.txt
# Vendsoft Courier Dependencies
playwright==1.41.0
requests==2.31.0
python-dotenv==1.0.0
